requirements.txt
certifi==2020.4.5.1
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.1.2
pytesseract==0.3.4
Werkzeug==1.0.1
